Update app.py
I polished the ceasar and rot13 cyphers. I started to explore a way to encrypt using the vigenere cypher and must admit it's real fun but lot of mind struggles. But I got the ascii code for the key and the message just need to merge them and transform them into characters.